Added the ability to move all cells down (and combine like ones)
Added two methods: moveEverythingDownward() and moveCellDownward();

The moveEverythingDownward method will move all the cells down and combine the cells that are equal. this method calls the moveCellDownward() method for every cell on the board

The moveCellDownward method will use recursion to move a cell as far down as possible. the cell will stop moving down under 2 conditions: either there is nothing below it or the number below it is a different number